last tile not displayed test

+++ src/carcassonne/view/GameWindow.java
@@ -299,13 +299,14 @@ public class GameWindow extends JFrame implements ViewWindow{
                         (int) (windowLocalY + tileSize * tile.getY()),
                         (int) tileSize, (int) tileSize, null);
             }
-/*
+
             for (Coordinates coordinates: possibleTileLocations) {
-                g.drawRect(windowLocalX + tileSize * coordinates.getX() + rectangleMargin,
-                           windowLocalY + tileSize * coordinates.getY()  + rectangleMargin,
-                           tileSize - 2 * rectangleMargin, tileSize - 2 * rectangleMargin);
+                g.drawRect((int) (windowLocalX + tileSize * coordinates.getX() + rectangleMargin),
+                        (int) (windowLocalY + tileSize * coordinates.getY()  + rectangleMargin),
+                        (int) (tileSize - 2 * rectangleMargin),
+                        (int) (tileSize - 2 * rectangleMargin));
             }
-*/
+
         }
     }
 }
